Agent_week5_memory_added/requirements.txt
openai>=1.40.0
chromadb>=0.5.5
pypdf>=4.2.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
tavily-python>=0.3.6
datasets>=2.20
transformers>=4.43
peft>=0.11
accelerate>=0.33
scikit-learn>=1.5
